+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Si vous souhaitez voir le détail, cliquez sur les numéros qui vous renverrons 
 # [2.14] - Correctifs divers et améliorations
 ## Améliorations :
 - [#519](https://github.com/Refhi/Weda-Helper/issues/519) - Ajout d’un champ de recherche pour filtrer les pharmacies dans la sélection de pharmacie lors de la création d’une prescription.
+- [#520](https://github.com/Refhi/Weda-Helper/issues/520) - Le champ de date d’expiration du vaccin est désormais automatiquement formaté.
 
 ## Fix :
 - [#517](https://github.com/Refhi/Weda-Helper/issues/517) - La lecture automatique de la carte vitale fonctionne de nouveau correctement pour les nouvelles installations.